cleanup parallel implementation
