+++ test/runtests.jl
@@ -1,5 +1,4 @@
 include("test_utils.jl")
 include("test_softmax.jl")
-
-addprocs(4)
 include("test_embedding.jl")
+#include("test_parallel.jl")
